+++ sentry_flutter_plus/lib/sentry_flutter_plus.dart
@@ -2,8 +2,8 @@ library sentry_flutter_plus;
 
 export 'src/extension.dart';
 export 'src/method_channel/sentry_binary_messenger.dart';
-export 'src/widget_tree_attachment.dart';
 export 'src/integrations/platform_menu_integration.dart'
     show PlatformMenuIntegration;
 export 'src/integrations/binding_integration.dart';
 export 'src/integrations/tree_walker_integration.dart';
+export 'src/integrations/jank_detection_integration.dart';