+++ sentry_flutter_plus/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0
+
+* Delete widget tree attachment
+* Remove ability to add screenshots to reports, [since Sentry is actively working on it](https://github.com/getsentry/sentry-dart/pull/1088)
+* Add jank detection integration: `JankDetectionIntegration`
+
 ## 0.1.0
 
 * Min Flutter version is now 3.3